// src/types/opportunity.ts
export interface ICreateOpportunity {
  name: string;
}

export interface IOpportunity {
  currency?: string;
  "@odata.etag": string;
  prioritycode: number;
  completeinternalreview: boolean;
  stepname: string;
  filedebrief: boolean;
  totaltax: number;
  estimatedclosedate: string;
  _pricelevelid_value: string;
  totalamount: number;
  confirminterest: boolean;
  captureproposalfeedback: boolean;
  exchangerate: number;
  opportunityid: string;
  _parentcontactid_value: string;
  identifycompetitors: boolean;
  _parentaccountid_value: string;
  decisionmaker: boolean;
  totallineitemamount: number;
  isrevenuesystemcalculated: boolean;
  modifiedon: string;
  _owninguser_value: string;
  skippricecalculation: number;
  presentproposal: boolean;
  proposedsolution: string;
  _ownerid_value: string;
  sendthankyounote: boolean;
  identifycustomercontacts: boolean;
  name: string;
  evaluatefit: boolean;
  totalamountlessfreight: number;
  totallineitemdiscountamount: number;
  totalamountlessfreight_base: number;
  totaldiscountamount: number;
  statuscode: number;
  createdon: string;
  versionnumber: number;
  emailaddress: string;
  msdyn_gdproptout: boolean;
  customerneed: string;
  _msdyn_predictivescoreid_value: string;
  totaltax_base: number;
  totallineitemamount_base: number;
  estimatedvalue: number;
  totalamount_base: number;
  developproposal: boolean;
  purchaseprocess: number;
  description: string;
  _msdyn_opportunitykpiid_value: string;
  resolvefeedback: boolean;
  _transactioncurrencyid_value: string;
  totaldiscountamount_base: number;
  estimatedvalue_base: number;
  msdyn_forecastcategory: number;
  _modifiedby_value: string;
  presentfinalproposal: boolean;
  _createdby_value: string;
  timezoneruleversionnumber: number;
  currentsituation: string;
  pricingerrorcode: number;
  salesstagecode: number;
  totallineitemdiscountamount_base: number;
  _modifiedonbehalfby_value: string;
  purchasetimeframe: number;
  identifypursuitteam: boolean;
  closeprobability: number;
  participatesinworkflow: boolean;
  statecode: number;
  _owningbusinessunit_value: string;
  pursuitdecision: boolean;
  opportunityratingcode: number;
  _customerid_value: string;
  completefinalproposal: boolean;
  msdyn_opportunityscore: null;
  msdyn_scorehistory: null;
  actualvalue_base: null;
  msdyn_opportunityscoretrend: null;
  _msdyn_segmentid_value: null;
  budgetstatus: null;
  _accountid_value: null;
  finaldecisiondate: null;
  _slainvokedid_value: null;
  msdyn_similaropportunities: null;
  msdyn_opportunitygrade: null;
  quotecomments: null;
  timeline: null;
  qualificationcomments: null;
  _contactid_value: null;
  _campaignid_value: null;
  stageid: null;
  lastonholdtime: null;
  need: null;
  stepid: null;
  freightamount_base: null;
  processid: null;
  discountamount_base: null;
  onholdtime: null;
  schedulefollowup_qualify: null;
  importsequencenumber: null;
  utcconversiontimezonecode: null;
  discountamount: null;
  salesstage: null;
  _createdonbehalfby_value: null;
  traversedpath: null;
  _originatingleadid_value: null;
  customerpainpoints: null;
  budgetamount: null;
  _owningteam_value: null;
  scheduleproposalmeeting: null;
  timespentbymeonemailandmeetings: null;
  teamsfollowed: null;
  overriddencreatedon: null;
  initialcommunication: null;
  schedulefollowup_prospect: null;
  msdyn_scorereasons: null;
  budgetamount_base: null;
  freightamount: null;
  actualclosedate: null;
  actualvalue: null;
  discountpercentage: null;
  _slaid_value: null;
}