// src/types/activity.ts
export interface IActivity {
  "@odata.type": string;
  "@odata.etag": string;
  scheduledstart: string;
  scheduleddurationminutes: number;
  _regardingobjectid_value: string;
  statuscode: number;
  leftvoicemail: boolean;
  createdon: string;
  statecode: number;
  isbilled: boolean;
  subject: string;
  _ownerid_value: string;
  modifiedon: string;
  versionnumber: number;
  prioritycode: number;
  timezoneruleversionnumber: number;
  isregularactivity: boolean;
  ismapiprivate: boolean;
  scheduledend: string;
  _modifiedby_value: string;
  activitytypecode: string;
  instancetypecode: number;
  isworkflowcreated: boolean;
  _createdby_value: string;
  _owningbusinessunit_value: string;
  actualdurationminutes: number;
  _owninguser_value: string;
  activityid: string;
  deliveryprioritycode: number;
  onholdtime: null;
  _slaid_value: null;
  seriesid: null;
  actualstart: null;
  exchangeitemid: null;
  _transactioncurrencyid_value: null;
  stageid: null;
  descriptionblobid: null;
  formattedscheduledend: null;
  postponeactivityprocessinguntil: null;
  senton: null;
  _serviceid_value: null;
  activityadditionalparams: null;
  _modifiedonbehalfby_value: null;
  exchangerate: null;
  formattedscheduledstart: null;
  actualend: null;
  descriptionblobid_name: null;
  community: null;
  _slainvokedid_value: null;
  utcconversiontimezonecode: null;
  processid: null;
  _createdonbehalfby_value: null;
  exchangeweblink: null;
  lastonholdtime: null;
  deliverylastattemptedon: null;
  _sendermailboxid_value: null;
  description: null;
  traversedpath: null;
  sortdate: null;
  _owningteam_value: null;
}