// src/types/call.ts
export interface ICreateCall {
  subject: string;
  scheduledstart: string;
  scheduledend: string;
  _regardingobjectid_value: string;
  _ownerid_value: string;
  description: string;
}

export interface ICall {
  "@odata.etag": string;
  scheduledstart: string;
  scheduleddurationminutes: number;
  _regardingobjectid_value: string;
  statuscode: number;
  description: string;
  leftvoicemail: boolean;
  createdon: string;
  statecode: number;
  isbilled: boolean;
  subject: string;
  directioncode: boolean;
  _ownerid_value: string;
  phonenumber: string;
  modifiedon: string;
  versionnumber: number;
  prioritycode: number;
  timezoneruleversionnumber: number;
  isregularactivity: boolean;
  scheduledend: string;
  _modifiedby_value: string;
  actualstart: string;
  activitytypecode: string;
  isworkflowcreated: boolean;
  _createdby_value: string;
  actualend: string;
  _owningbusinessunit_value: string;
  _owninguser_value: string;
  activityid: string;
  actualdurationminutes: number;
  formattedscheduledend: null;
  senton: null;
  exchangeweblink: null;
  _slainvokedid_value: null;
  stageid: null;
  lastonholdtime: null;
  ismapiprivate: null;
  sortdate: null;
  _modifiedonbehalfby_value: null;
  _sendermailboxid_value: null;
  processid: null;
  onholdtime: null;
  postponeactivityprocessinguntil: null;
  utcconversiontimezonecode: null;
  descriptionblobid_name: null;
  traversedpath: null;
  _createdonbehalfby_value: null;
  activityadditionalparams: null;
  community: null;
  _owningteam_value: null;
  formattedscheduledstart: null;
  deliveryprioritycode: null;
  exchangerate: null;
  exchangeitemid: null;
  _transactioncurrencyid_value: null;
  deliverylastattemptedon: null;
  descriptionblobid: null;
  seriesid: null;
  _slaid_value: null;
  _serviceid_value: null;
  instancetypecode: null;
  overriddencreatedon: null;
  subcategory: null;
  category: null;
  importsequencenumber: null;
  subscriptionid: null;
}