// src/types/appointment.ts
export interface ICreateAppointment {
  subject: string;
  scheduledstart: string;
  scheduledend: string;
  _regardingobjectid_value: string;
  _ownerid_value: string;
  description: string;
}

export interface IAppointment {
  "@odata.etag": string;
  scheduledstart: string;
  scheduleddurationminutes: number;
  _regardingobjectid_value: string;
  statuscode: number;
  activitytypecode: string;
  createdon: string;
  statecode: number;
  isbilled: boolean;
  subject: string;
  _ownerid_value: string;
  isunsafe: number;
  modifiedon: string;
  versionnumber: number;
  prioritycode: number;
  isdraft: boolean;
  timezoneruleversionnumber: number;
  isregularactivity: boolean;
  isalldayevent: boolean;
  ismapiprivate: boolean;
  scheduledend: string;
  _modifiedby_value: string;
  actualstart: string;
  isonlinemeeting: boolean;
  instancetypecode: number;
  activityadditionalparams: string;
  isworkflowcreated: boolean;
  _createdby_value: string;
  attachmenterrors: number;
  actualend: string;
  _owningbusinessunit_value: string;
  _owninguser_value: string;
  activityid: string;
  actualdurationminutes: number;
  formattedscheduledend: null;
  senton: null;
  exchangeweblink: null;
  _slainvokedid_value: null;
  leftvoicemail: null;
  stageid: null;
  lastonholdtime: null;
  sortdate: null;
  _modifiedonbehalfby_value: null;
  _sendermailboxid_value: null;
  description: null;
  processid: null;
  onholdtime: null;
  postponeactivityprocessinguntil: null;
  utcconversiontimezonecode: null;
  descriptionblobid_name: null;
  traversedpath: null;
  _createdonbehalfby_value: null;
  community: null;
  _owningteam_value: null;
  formattedscheduledstart: null;
  deliveryprioritycode: null;
  exchangerate: null;
  exchangeitemid: null;
  _transactioncurrencyid_value: null;
  deliverylastattemptedon: null;
  descriptionblobid: null;
  seriesid: null;
  _slaid_value: null;
  _serviceid_value: null;
  outlookownerapptid: null;
  globalobjectid: null;
  overriddencreatedon: null;
  subcategory: null;
  category: null;
  onlinemeetingchatid: null;
  importsequencenumber: null;
  onlinemeetingjoinurl: null;
  onlinemeetingid: null;
  originalstartdate: null;
  location: null;
  subscriptionid: null;
  onlinemeetingtype: null;
  modifiedfieldsmask: null;
}